cbn_repo/requirements.txt
Cython==0.29
pymoo==0.4.1
timm==0.1.30
pySOT==0.2.3
pydacefit==1.0.1
filelock==3.4.1
torchprofile==0.0.1